+++ src/compat/_internal/toKey.ts
@@ -1,5 +1,3 @@
-import { isSymbol } from '../predicate/isSymbol';
-
 /**
  * Converts `value` to a string key if it's not a string or symbol.
  *
@@ -7,14 +5,9 @@ import { isSymbol } from '../predicate/isSymbol';
  * @param {*} value The value to inspect.
  * @returns {string|symbol} Returns the key.
  */
-export function toKey(value: unknown) {
-  if (typeof value === 'string' || isSymbol(value)) {
-    return value;
-  }
-
-  if (Object.is(value?.valueOf(), -0)) {
+export function toKey(value: number) {
+  if (Object.is(value, -0)) {
     return '-0';
   }
-
-  return `${value}`;
+  return value.toString();
 }